+++ Locatics/LocaticsTests/App/Coordinators/ApplicationCoordinatorTests.swift
@@ -85,54 +85,3 @@ class ApplicationCoordinatorTests: XCTestCase {
         XCTAssertFalse(mockOnboardingCoordinator.startWasCalled)
     }
 }
-
-private extension ApplicationCoordinatorTests {
-    class MockCoordinatorFactory: CoordinatorFactoryInterface {
-
-        let onboardingCoordinator: MockOnboardingCoordinator
-        let mainCoordinator: MockMainCoordinator
-
-        init(onboardingCoordinator: MockOnboardingCoordinator, mainCoordinator: MockMainCoordinator) {
-            self.onboardingCoordinator = onboardingCoordinator
-            self.mainCoordinator = mainCoordinator
-        }
-
-        func createOnboardingFlow(root: UINavigationController) -> CoordinatorInterface & OnboardingCoordinatorOutput {
-            return onboardingCoordinator
-        }
-
-        func createMainFlow(root: TabBarControllerInterface) -> CoordinatorInterface {
-            return mainCoordinator
-        }
-
-        func createMapFlow(root: UINavigationController) -> CoordinatorInterface {
-            return LocaticsMapCoordinator()
-        }
-
-        func createLocaticsFlow(root: UINavigationController) -> CoordinatorInterface {
-            return LocaticsCoordinator()
-        }
-    }
-
-    class MockOnboardingCoordinator: CoordinatorInterface, OnboardingCoordinatorOutput {
-        var finishedOnboarding: (() -> Void)?
-
-        var startWasCalled = false
-
-        func start() {
-            startWasCalled = true
-        }
-
-        func triggerFinishOnboarding() {
-            finishedOnboarding?()
-        }
-    }
-
-    class MockMainCoordinator: CoordinatorInterface {
-        var startWasCalled = false
-
-        func start() {
-            startWasCalled = true
-        }
-    }
-}
